refactor(zudo-doc): split note tray index renderers
